only use esm migration provider in windows tests
